+++ backend/src/main/kotlin/com/pao/plugins/Routing.kt
@@ -6,6 +6,7 @@ import com.pao.routes.otherMessage
 
 // imports essenciais do routing
 import io.ktor.server.application.*
+import io.ktor.server.http.content.*
 import io.ktor.server.routing.*
 
 import io.ktor.server.request.*
@@ -16,5 +17,10 @@ fun Application.configureRouting() {
     routing {
         someMessage()
         otherMessage()
+
+        // garante pegar as imagens de uma pasta
+        static{
+            resources("static")
+        }
     }
 }
