+++ frontend/app/src/main/AndroidManifest.xml
@@ -31,16 +31,16 @@
         android:testOnly="true"
         android:theme="@style/Theme.Paozim">
         <activity
-            android:name=".HomeActivity"
+            android:name=".activities.HomeActivity"
             android:exported="false" />
         <activity
-            android:name=".BackendActivity"
+            android:name=".activities.BackendActivity"
             android:exported="false" />
         <activity
-            android:name=".LoginActivity"
+            android:name=".activities.LoginActivity"
             android:exported="false" />
         <activity
-            android:name=".MainActivity"
+            android:name=".activities.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
