+++ frontend/app/src/main/java/com/mobile/paozim/activities/LoginActivity.kt
@@ -1,24 +1,17 @@
-package com.mobile.paozim
+package com.mobile.paozim.activities
 
 import android.content.Intent
 import android.os.Bundle
 import android.util.Log
 import android.widget.Toast
-import androidx.activity.enableEdgeToEdge
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.ViewCompat
-import androidx.core.view.WindowInsetsCompat
 import com.mobile.paozim.databinding.ActivityLoginBinding
 import com.mobile.paozim.testdata.RetrofitClient
 import com.mobile.paozim.testdata.UserInfo
-import com.mobile.paozim.testdata.UsersAPI
 
 import retrofit2.Call
 import retrofit2.Callback
-import retrofit2.HttpException
 import retrofit2.Response
-import retrofit2.create
-import java.io.IOException
 import kotlin.properties.Delegates
 
 class LoginActivity : AppCompatActivity() {
