+++ frontend/app/src/main/java/com/mobile/paozim/fragments/ProfileFragment.kt
@@ -1,15 +1,11 @@
-package com.mobile.paozim
+package com.mobile.paozim.fragments
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-
-// TODO: Rename parameter arguments, choose names that match
-// the fragment initialization parameters, e.g. ARG_ITEM_NUMBER
-private const val ARG_PARAM1 = "param1"
-private const val ARG_PARAM2 = "param2"
+import androidx.fragment.app.Fragment
+import com.mobile.paozim.R
 
 /**
  * A simple [Fragment] subclass.
@@ -24,8 +20,8 @@ class ProfileFragment : Fragment() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         arguments?.let {
-            param1 = it.getString(ARG_PARAM1)
-            param2 = it.getString(ARG_PARAM2)
+            param1 = it.getString("bla1")
+            param2 = it.getString("bla2")
         }
     }
 
@@ -51,8 +47,8 @@ class ProfileFragment : Fragment() {
         fun newInstance(param1: String, param2: String) =
             ProfileFragment().apply {
                 arguments = Bundle().apply {
-                    putString(ARG_PARAM1, param1)
-                    putString(ARG_PARAM2, param2)
+                    putString("bla1", param1)
+                    putString("bla2", param2)
                 }
             }
     }
