+++ frontend/app/src/main/res/layout/activity_home.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".HomeActivity">
+    tools:context=".activities.HomeActivity">
 
     <fragment
         android:id="@+id/host_fragment"
