+++ frontend/app/src/main/res/navigation/nav_graph.xml
@@ -7,17 +7,17 @@
 
     <fragment
         android:id="@+id/homeFragment"
-        android:name="com.mobile.paozim.HomeFragment"
+        android:name="com.mobile.paozim.fragments.HomeFragment"
         android:label="fragment_home"
         tools:layout="@layout/fragment_home" />
     <fragment
         android:id="@+id/membershipFragment"
-        android:name="com.mobile.paozim.MembershipFragment"
+        android:name="com.mobile.paozim.fragments.MembershipFragment"
         android:label="fragment_membership"
         tools:layout="@layout/fragment_membership" />
     <fragment
         android:id="@+id/profileFragment"
-        android:name="com.mobile.paozim.ProfileFragment"
+        android:name="com.mobile.paozim.fragments.ProfileFragment"
         android:label="fragment_profile"
         tools:layout="@layout/fragment_profile" />
 </navigation>
